+++ library/src/main/java/com/google/android/exoplayer/chunk/ContainerMediaChunk.java
@@ -167,7 +167,7 @@ public class ContainerMediaChunk extends BaseMediaChunk implements SingleTrackOu
         bytesLoaded = (int) (input.getPosition() - dataSpec.absoluteStreamPosition);
       }
     } finally {
-      dataSource.close();
+      Util.closeQuietly(dataSource);
     }
   }